windows-once: Fix race (regression yesterday).
* lib/windows-once.h (glwthread_once_t): Change type of inited to LONG.
* lib/windows-once.c (glwthread_once): Increment inited from 1 to 2, to
ensure that DeleteCriticalSection gets invoked only once.

+++ lib/windows-once.h
@@ -25,7 +25,7 @@
 
 typedef struct
         {
-          volatile int inited;
+          volatile LONG inited;
           volatile LONG num_threads;
           volatile LONG started;
           CRITICAL_SECTION lock;
